#22 Architectural improvement: Unify AnkiConnect client and shared parsers
- Move AnkiConnectClient base class to @ankiniki/shared
- Relocate import parsers and tests from backend to @ankiniki/shared
- Standardize server port and URL constants in @ankiniki/shared
- Update CLI and Backend to use unified components and constants
