+++ packages/shared/src/__tests__/import-parsers.test.ts
@@ -8,7 +8,7 @@ import {
   CsvImportOptionsSchema,
   JsonImportOptionsSchema,
   MarkdownImportOptionsSchema,
-} from '../lib/import-parsers';
+} from '../import-parsers';
 
 // ─── parseFrontmatter ─────────────────────────────────────────────────────────
